kernel: File cleanup pt1
