--- src/core/libraries/kernel/event_flag/event_flag.h
@@ -1,40 +1,0 @@
-// SPDX-FileCopyrightText: Copyright 2024 shadPS4 Emulator Project
-// SPDX-License-Identifier: GPL-2.0-or-later
-
-#pragma once
-
-#include "common/types.h"
-#include "event_flag_codes.h"
-#include "event_flag_obj.h"
-
-namespace Core::Loader {
-class SymbolsResolver;
-}
-
-namespace Libraries::Kernel {
-
-using OrbisKernelUseconds = u32;
-using OrbisKernelEventFlag = EventFlagInternal*;
-
-struct OrbisKernelEventFlagOptParam {
-    size_t size;
-};
-
-int PS4_SYSV_ABI sceKernelCreateEventFlag(OrbisKernelEventFlag* ef, const char* pName, u32 attr,
-                                          u64 initPattern,
-                                          const OrbisKernelEventFlagOptParam* pOptParam);
-int PS4_SYSV_ABI sceKernelDeleteEventFlag(OrbisKernelEventFlag ef);
-int PS4_SYSV_ABI sceKernelOpenEventFlag();
-int PS4_SYSV_ABI sceKernelCloseEventFlag();
-int PS4_SYSV_ABI sceKernelClearEventFlag(OrbisKernelEventFlag ef, u64 bitPattern);
-int PS4_SYSV_ABI sceKernelCancelEventFlag(OrbisKernelEventFlag ef, u64 setPattern,
-                                          int* pNumWaitThreads);
-int PS4_SYSV_ABI sceKernelSetEventFlag(OrbisKernelEventFlag ef, u64 bitPattern);
-int PS4_SYSV_ABI sceKernelPollEventFlag(OrbisKernelEventFlag ef, u64 bitPattern, u32 waitMode,
-                                        u64* pResultPat);
-int PS4_SYSV_ABI sceKernelWaitEventFlag(OrbisKernelEventFlag ef, u64 bitPattern, u32 waitMode,
-                                        u64* pResultPat, OrbisKernelUseconds* pTimeout);
-
-void RegisterKernelEventFlag(Core::Loader::SymbolsResolver* sym);
-
-} // namespace Libraries::Kernel
