--- src/core/libraries/kernel/event_flag/event_flag_codes.h
@@ -1,14 +1,0 @@
-// SPDX-FileCopyrightText: Copyright 2024 shadPS4 Emulator Project
-// SPDX-License-Identifier: GPL-2.0-or-later
-
-#pragma once
-
-constexpr int ORBIS_KERNEL_EVF_ATTR_TH_FIFO = 0x01;
-constexpr int ORBIS_KERNEL_EVF_ATTR_TH_PRIO = 0x02;
-constexpr int ORBIS_KERNEL_EVF_ATTR_SINGLE = 0x10;
-constexpr int ORBIS_KERNEL_EVF_ATTR_MULTI = 0x20;
-
-constexpr int ORBIS_KERNEL_EVF_WAITMODE_AND = 0x01;
-constexpr int ORBIS_KERNEL_EVF_WAITMODE_OR = 0x02;
-constexpr int ORBIS_KERNEL_EVF_WAITMODE_CLEAR_ALL = 0x10;
-constexpr int ORBIS_KERNEL_EVF_WAITMODE_CLEAR_PAT = 0x20;
