+++ src/core/libraries/kernel/kernel.h
