+++ src/core/libraries/kernel/threads/thr_attr.cpp
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: GPL-2.0-or-later
 #pragma clang optimize off
 #include "core/libraries/error_codes.h"
-#include "core/libraries/kernel/libkernel.h"
+#include "core/libraries/kernel/kernel.h"
 #include "core/libraries/kernel/threads/thread_state.h"
 #include "core/libraries/kernel/threads/threads.h"
 #include "core/libraries/libs.h"
