+++ src/core/libraries/kernel/threads/thr_cond.cpp
@@ -4,7 +4,7 @@
 #include <cstring>
 #include "common/assert.h"
 #include "core/libraries/error_codes.h"
-#include "core/libraries/kernel/libkernel.h"
+#include "core/libraries/kernel/kernel.h"
 #include "core/libraries/kernel/threads/threads.h"
 #include "core/libraries/libs.h"
