+++ src/core/libraries/kernel/threads/thr_create.cpp
@@ -5,7 +5,7 @@
 #include "common/thread.h"
 #include "core/debug_state.h"
 #include "core/libraries/error_codes.h"
-#include "core/libraries/kernel/libkernel.h"
+#include "core/libraries/kernel/kernel.h"
 #include "core/libraries/kernel/threads/thread_state.h"
 #include "core/libraries/kernel/threads/threads.h"
 #include "core/libraries/libs.h"
