+++ src/core/libraries/kernel/threads/thr_mutex.cpp
@@ -5,7 +5,7 @@
 #include "common/scope_exit.h"
 #include "common/types.h"
 #include "core/libraries/error_codes.h"
-#include "core/libraries/kernel/libkernel.h"
+#include "core/libraries/kernel/kernel.h"
 #include "core/libraries/kernel/threads/threads.h"
 #include "core/libraries/libs.h"
