+++ src/core/libraries/kernel/threads/thr_sem.cpp
@@ -7,7 +7,7 @@
 #include "common/assert.h"
 #include "common/logging/log.h"
 #include "core/libraries/error_codes.h"
-#include "core/libraries/kernel/libkernel.h"
+#include "core/libraries/kernel/kernel.h"
 #include "core/libraries/kernel/time_management.h"
 #include "core/libraries/libs.h"
